Add simple script to generate keys needed for relayer
This removes the need to have the relayer magically auto generate the
keys when it spins up for the first time. The workflow for doing ops on
a new relayer environment is now that you would generate the keys on a
dev box, put those keys in 1pwd, and then set those values in the
necessary fly secrets, and then the relayer will read them when it spins
up.